Rename dateTimeWithSetLocale.js to frenchDateTime.js #250
